MDL-32205 Introduce support for course format data conversion
The new core moodle1_course_format_handler acts as a communication
bridge between the core and course format plugins. Its main
responsibility is to collect the list of paths with FORMATDATA that
installed plugins want to observer. It the attaches itself as the
observer of these paths. As moodle.xml is being parsed, the
moodle1_course_format_handler re-dispatches grouped FORMATDATA contents
to the appropriate course format plugin for conversion.

Course format plugins are supposed to provide a subclass of
moodle1_format_handler with the required methods implemented.

Data converted from the FORMATDATA element can be attached either to the
course structure or individual section structures.