MDL-49619 workshop: Add support for method calls delegation
At certain places, the workshop now gives its subplugins a chance to
intercept the output and/or data processing. These places look like

    $workshop->delegate->foobar($params);

All subplugins that provide the method foobar() in a dedicated class are
then checked and the method is called. The order in which the subplugins
are checked is not guaranteed and subplugins should not rely on it.

The subplugin will typically either modify objects passed as
parameters or output directly.