MDL-49619 workshop: Reflect the issue peer-review comments
This patch implements some of the suggestions raised by Tim Hunt during
the peer-review of the issue.

The workshoptools subplugins are now marked as uninstallable all the
time. There is no need for the exception for the standard ones.

The delegate base class now implements all the supported delegable
methods as empty. We are aware this leads to the partial duplication of
the code but there were performance benefits discussed and agreed. See
the discussion in the tracker for details.