MDL-50446 block_site_main_menu: Fix the onclick property handling
The modules can set the onclick property to any valid JS code, not only
the single function name. The patch wraps the provided code with an
anonymous wrapper.

Note that ideally we should halt the event propagation in the same way
as navigationlib.php does it (grep for legacy_activity_onclick_handler
there). But this seems to work, too.