MDL-52574 get_role_users() fails with specific fields on PostgreSQL
The function get_role_users() requires the caller to include the $sort fields in
the $fields argument. On PostgreSQL this will cause the function to fail when
the default $sort fields aren't part of the requested fields. The behavior of
the function is augmented to add the $sort fields to $fields if they are not
already present.

Signed-off-by: Alexandru Elisei <alexandru.elisei@gmail.com>