MDL-55484 calendar: Fix incorrect date of exported all day events
When exporting a calendar into iCalendar format, events with no duration
are interpreted as all day events. To set the correct date for such an
event, we must not use the raw UTC timestamp of the event. Instead we
need to look back and guess what actual date the user had seen in the
date/time picker while creating the event.