MDL-60510 setup: Do not check lib.php for callbacks when needs upgrade
If the Moodle needs upgrading (such as a new plugin code is deployed but
the database upgrade has not run yet), it is not safe to load plugins'
lib.php files - there can be unsatisfied dependencies.

The patch introduces a new optional flag so that we can easily check if
these early callbacks should be checked or not.