MDL-81183 admin: Use version and branch number from version.php
When updating the available plugin updates together with the core, we
cannot rely on the core version and branch values stored in the CFG
object. Those describe the database version, not the on-disk version we
are going to upgrade to.

We need to read and use the actual value from the version.php without
any caching.